Refactor and reorganize helpers and modules
