+++ src/contents/flights.ts
@@ -1,7 +1,7 @@
 import type { PlasmoCSConfig } from "plasmo"
 
 import { loadLinksTable } from "~/modules/loadLinksTable"
-import { toggleLinksTableSortOrder } from "~helpers/utils"
+import { updateLinksTable } from "~/modules/updateLinksTable"
 
 export const config: PlasmoCSConfig = {
   world: "MAIN",
@@ -10,7 +10,30 @@ export const config: PlasmoCSConfig = {
 }
 
 window.addEventListener("DOMContentLoaded", async () => {
-  window.toggleLinksTableSortOrder = toggleLinksTableSortOrder
+  window.toggleLinksTableSortOrder = function toggleLinksTableSortOrder(sortHeader: HTMLElement): void {
+    const sortOrder = sortHeader.getAttribute("data-sort-order")
+
+    if (sortOrder === "ascending") {
+      sortHeader.setAttribute("data-sort-order", "descending")
+    } else {
+      sortHeader.setAttribute("data-sort-order", "ascending")
+    }
+
+    Array.from(sortHeader.parentElement?.children ?? []).forEach((sibling) => {
+      if (sibling !== sortHeader) {
+        sibling.classList.remove("selected")
+      }
+    })
+
+    sortHeader.classList.add("selected")
+
+    const sortProperty = sortHeader.getAttribute("data-sort-property")
+    const updatedSortOrder = sortHeader.getAttribute("data-sort-order")
+
+    if (sortProperty && updatedSortOrder) {
+      updateLinksTable(sortProperty, updatedSortOrder)
+    }
+  }
   window.loadLinksTable = loadLinksTable
 })
