+++ src/modules/updateLinksTable.ts
@@ -1,6 +1,7 @@
 import { appendCell, prependCell } from "~/helpers/tables"
-import { getStyleFromTier, getTierFromPercent } from "~/helpers/tiers"
-import { getShortModelName, selectLinkFromTable } from "~/helpers/utils"
+import { selectLinkFromTable } from "~/helpers/tables"
+import { getShortModelName } from "~/helpers/tables"
+import { getStyleFromTier, getTierFromPercent } from "~/helpers/utils"
 
 export function updateLinksTable(sortProperty: string, sortOrder: string) {
   let colorKeyMaps: { [key: string]: WeakMap<Link, any> } = {}
